Improve readability (hopefully), re-add permissions
