+++ src/main/java/de/t14d3/zones/brigadier/SubCommandArgument.java
@@ -2,15 +2,9 @@ package de.t14d3.zones.brigadier;
 
 import com.mojang.brigadier.arguments.ArgumentType;
 import com.mojang.brigadier.arguments.StringArgumentType;
-import com.mojang.brigadier.context.CommandContext;
 import com.mojang.brigadier.exceptions.CommandSyntaxException;
-import com.mojang.brigadier.suggestion.Suggestions;
-import com.mojang.brigadier.suggestion.SuggestionsBuilder;
-import io.papermc.paper.command.brigadier.MessageComponentSerializer;
 import io.papermc.paper.command.brigadier.argument.CustomArgumentType;
 import org.jetbrains.annotations.NotNull;
-
-import java.util.concurrent.CompletableFuture;
 
 
 @SuppressWarnings("UnstableApiUsage")
@@ -23,16 +17,5 @@ public class SubCommandArgument implements CustomArgumentType.Converted<SubComma
     @Override
     public @NotNull ArgumentType<String> getNativeType() {
         return StringArgumentType.string();
-    }
-
-    @Override
-    public <S> @NotNull CompletableFuture<Suggestions> listSuggestions(@NotNull CommandContext<S> context, @NotNull SuggestionsBuilder builder) {
-        for (SubCommands subCommand : SubCommands.values()) {
-            String[] args = context.getInput().split(" ");
-            if (args.length == 1 || subCommand.name().toLowerCase().startsWith(args[1].toLowerCase())) {
-                builder.suggest(subCommand.name().toLowerCase(), MessageComponentSerializer.message().serialize(subCommand.getInfo()));
-            }
-        }
-        return builder.buildFuture();
     }
 }
