+++ src/main/resources/paper-plugin.yml
@@ -3,3 +3,93 @@ version: '${version}'
 main: de.t14d3.zones.Zones
 api-version: '1.21'
 bootstrapper: de.t14d3.zones.PaperBootstrap
+permissions:
+  zones.info:
+    description: Allow players to view region information
+    default: true
+    children:
+      zones.info.other: true
+  zones.delete:
+    description: Allow players to delete regions
+    default: true
+    children:
+      zones.delete.other: true
+  zones.create:
+    description: Allow players to create regions
+    default: true
+    children:
+      zones.create.overlap: true
+  zones.subcreate:
+    description: Allow players to create sub-regions
+    default: true
+    children:
+      zones.subcreate.other: true
+  zones.cancel:
+    description: Allow players to cancel the current operation
+    default: true
+  zones.list:
+    description: Allow players to list all regions
+    default: true
+  zones.set:
+    description: Allow players to set permissions for regions
+    default: true
+    children:
+      zones.set.other: true
+  zones.expand:
+    description: Allow players to expand regions
+    default: true
+    children:
+      zones.expand.other: true
+      zones.expand.overlap: true
+  zones.select:
+    description: Allow players to select regions
+    default: true
+    children:
+      zones.select.other: true
+  zones.rename:
+    description: Allow players to rename regions
+    default: true
+    children:
+      zones.rename.other: true
+
+  zones.save:
+    description: Allow players to save all regions to file
+    default: op
+  zones.load:
+    description: Allow players to load all regions from file
+    default: op
+
+  zones.info.other:
+    description: Allow players to view other player's regions
+    default: op
+  zones.select.other:
+    description: Allow players to select other player's regions
+    default: op
+  zones.delete.other:
+    description: Allow players to delete other player's regions
+    default: op
+  zones.subcreate.other:
+    description: Allow players to create sub-regions for other player's regions
+    default: op
+  zones.expand.other:
+    description: Allow players to expand regions over other regions
+    default: op
+  zones.expand.overlap:
+    description: Allow players to expand regions over other regions
+    default: op
+  zones.set.other:
+    description: Allow players to set permissions for other player's regions
+    default: op
+  zones.create.overlap:
+    description: Allow players to create regions that overlap existing regions
+    default: op
+  zones.rename.other:
+    description: Allow players to rename other player's regions
+    default: op
+
+  zones.bypass.unclaimed:
+    description: Allow players to interact with unclaimed regions
+    default: op
+  zones.bypass.claimed:
+    description: Allow players to interact with other players' regions
+    default: op
